+++ fastapi_limiter/__init__.py
@@ -3,7 +3,7 @@ from typing import Callable
 import aioredis
 from fastapi import HTTPException
 from starlette.requests import Request
-from starlette.status import HTTP_403_FORBIDDEN
+from starlette.status import HTTP_429_TOO_MANY_REQUESTS
 
 
 async def default_identifier(request: Request):
@@ -13,8 +13,16 @@ async def default_identifier(request: Request):
     return request.client.host
 
 
-async def default_callback(request: Request):
-    raise HTTPException(HTTP_403_FORBIDDEN, "The request is frequent")
+async def default_callback(request: Request, expire: int):
+    """
+    default callback when too many requests
+    :param request:
+    :param expire: The remaining seconds
+    :return:
+    """
+    raise HTTPException(
+        HTTP_429_TOO_MANY_REQUESTS, "Too Many Requests", headers={"Retry-After": str(expire)}
+    )
 
 
 class FastAPILimiter: