+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # ChangeLog
+
+## 0.1
+
+### 0.1.1
+
+- Configuring the global default through the FastAPILimiter.init method.
+- Update status to 429 when too many requests.
+- Update default_callback params and add `Retry-After` response header.